test: refactor rfc3339 serialization integration tests
