+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/webapp/api/v1/dao/IncidentDaoIT.java
@@ -11,6 +11,7 @@ import static io.camunda.operate.schema.indices.IndexDescriptor.DEFAULT_TENANT_I
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import io.camunda.operate.connect.OperateDateTimeFormatter;
 import io.camunda.operate.entities.ErrorType;
 import io.camunda.operate.entities.IncidentEntity;
 import io.camunda.operate.entities.IncidentState;
@@ -20,14 +21,18 @@ import io.camunda.operate.webapp.api.v1.entities.Incident;
 import io.camunda.operate.webapp.api.v1.entities.Query;
 import io.camunda.operate.webapp.api.v1.entities.Results;
 import io.camunda.operate.webapp.api.v1.exceptions.ResourceNotFoundException;
-import java.time.OffsetDateTime;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class IncidentDaoIT extends OperateSearchAbstractIT {
+  private final String firstIncidentCreationTime = "2024-02-15T22:40:10.834+0000";
+  private final String secondIncidentCreationTime = "2024-02-15T22:41:10.834+0000";
+
   @Autowired private IncidentDao dao;
 
   @Autowired private IncidentTemplate incidentIndex;
+
+  @Autowired private OperateDateTimeFormatter dateTimeFormatter;
 
   @Override
   protected void runAdditionalBeforeAllSetup() throws Exception {
@@ -41,7 +46,7 @@ public class IncidentDaoIT extends OperateSearchAbstractIT {
             .setState(IncidentState.ACTIVE)
             .setErrorMessage("Some error")
             .setTenantId(DEFAULT_TENANT_ID)
-            .setCreationTime(OffsetDateTime.now())
+            .setCreationTime(dateTimeFormatter.parseGeneralDateTime(firstIncidentCreationTime))
             .setJobKey(2251799813685260L));
 
     testSearchRepository.createOrUpdateDocumentFromObject(
@@ -54,7 +59,7 @@ public class IncidentDaoIT extends OperateSearchAbstractIT {
             .setState(IncidentState.ACTIVE)
             .setErrorMessage("Another error")
             .setTenantId(DEFAULT_TENANT_ID)
-            .setCreationTime(OffsetDateTime.now())
+            .setCreationTime(dateTimeFormatter.parseGeneralDateTime(secondIncidentCreationTime))
             .setJobKey(3251799813685260L));
 
     searchContainerManager.refreshIndices("*operate-incident*");
@@ -220,18 +225,66 @@ public class IncidentDaoIT extends OperateSearchAbstractIT {
     assertThat(checkIncident).isNotNull();
     assertThat(checkIncident)
         .extracting(
-            "processDefinitionKey", "processInstanceKey", "type", "state", "message", "tenantId")
+            "processDefinitionKey",
+            "processInstanceKey",
+            "type",
+            "state",
+            "message",
+            "tenantId",
+            "creationTime")
         .containsExactly(
             5147483648L,
             6147483648L,
             "JOB_NO_RETRIES",
             "ACTIVE",
             "Another error",
-            DEFAULT_TENANT_ID);
+            DEFAULT_TENANT_ID,
+            secondIncidentCreationTime);
   }
 
   @Test
   public void shouldThrowWhenKeyNotExists() {
     assertThrows(ResourceNotFoundException.class, () -> dao.byKey(1L));
+  }
+
+  @Test
+  public void shouldFilterByCreationDate() {
+    final Results<Incident> flowNodeInstanceResults =
+        dao.search(
+            new Query<Incident>()
+                .setFilter(new Incident().setCreationTime(firstIncidentCreationTime)));
+
+    assertThat(flowNodeInstanceResults.getTotal()).isEqualTo(1L);
+    assertThat(flowNodeInstanceResults.getItems().get(0).getCreationTime())
+        .isEqualTo(firstIncidentCreationTime);
+    assertThat(flowNodeInstanceResults.getItems().get(0).getMessage()).isEqualTo("Some error");
+  }
+
+  @Test
+  public void shouldFilterByCreationDateWithDateMath() {
+    final Results<Incident> incidentResults =
+        dao.search(
+            new Query<Incident>()
+                .setFilter(new Incident().setCreationTime(firstIncidentCreationTime + "||/d")));
+
+    assertThat(incidentResults.getTotal()).isEqualTo(2L);
+
+    Incident checkIncident =
+        incidentResults.getItems().stream()
+            .filter(item -> "Some error".equals(item.getMessage()))
+            .findFirst()
+            .orElse(null);
+    assertThat(checkIncident)
+        .extracting("creationTime", "message")
+        .containsExactly(firstIncidentCreationTime, "Some error");
+
+    checkIncident =
+        incidentResults.getItems().stream()
+            .filter(item -> "Another error".equals(item.getMessage()))
+            .findFirst()
+            .orElse(null);
+    assertThat(checkIncident)
+        .extracting("creationTime", "message")
+        .containsExactly(secondIncidentCreationTime, "Another error");
   }
 }
